requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.2.1

# API
fastapi==0.111.0
uvicorn==0.30.1

# Data & Utilities
numpy==1.26.4
pandas==2.2.2

# Geospatial
shapely==2.0.4
geojson==3.1.0

# Graph & Routing
networkx==3.3

# LLM & Agents
openai==1.35.3
langchain==0.2.6
langchain-openai==0.1.8
langgraph==0.0.62

# Dashboard
streamlit==1.35.0
streamlit-folium==0.21.0
folium==0.17.0

# CLI & Dev
tqdm==4.66.4

# Testing
pytest==8.2.2

sqlalchemy==2.0.30
psycopg2-binary==2.9.9

scikit-learn==1.5.1